+++ runtime/options.go
@@ -99,6 +99,12 @@ func WithMessageBudget(maxMessages int) Option {
 	}
 }
 
+func WithTokenBudget(maxTokens int) Option {
+	return func(a *Agent) {
+		a.sessionOptions = append(a.sessionOptions, conversation.WithTokenBudget(maxTokens))
+	}
+}
+
 func WithStream(stream bool) Option {
 	return func(a *Agent) { a.request.Stream = stream }
 }